Feat: add customer edit/delete post (#113)
* fix: move venueID from uuid to int type

* feat: add swagger generated client

* feat: add update method for customer posts

* feat: implement post editing and status management with database migration

* feat: integrate image handling to posts

* feat: add delete method for posts

* feat: integrate comments and likes to existing posts

* feat: add swagger documentation to posts

* feat: add descriptions for swagger doc status codes

* feat: add avatar and username to post api response

* fix: content type error hanlding

+++ migrations/20260326141321_add_posts.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS guest.posts (
     id BIGSERIAL PRIMARY KEY,
     customer_id UUID NOT NULL REFERENCES guest.customers(id) ON DELETE CASCADE,
 
-    venue_id INT NOT NULL,
+    venue_id BIGINT NOT NULL,
     text TEXT NOT NULL,
     rating SMALLINT NOT NULL,
     status VARCHAR(20) NOT NULL DEFAULT 'published',
